Move ProductTemplate type
