+++ packages/js/product-editor/src/blocks/product-fields/product-details-section-description/edit.tsx
@@ -29,14 +29,13 @@ import { useEntityId } from '@wordpress/core-data';
  * Internal dependencies
  */
 import { ProductEditorSettings } from '../../../components';
-import { ProductTemplate } from '../../../components/editor';
 import { BlockFill } from '../../../components/block-slot-fill';
 import { useValidations } from '../../../contexts/validation-context';
 import {
 	WPError,
 	getProductErrorMessage,
 } from '../../../utils/get-product-error-message';
-import { ProductEditorBlockEditProps } from '../../../types';
+import { ProductEditorBlockEditProps, ProductTemplate } from '../../../types';
 import { ProductDetailsSectionDescriptionBlockAttributes } from './types';
 
 export function ProductDetailsSectionDescriptionBlockEdit( {
