+++ packages/js/product-editor/src/components/editor/types.ts
@@ -8,22 +8,17 @@ import {
 } from '@wordpress/block-editor';
 import { Template } from '@wordpress/blocks';
 
+/**
+ * Internal dependencies
+ */
+import { ProductTemplate } from '../../types';
+
 export type LayoutTemplate = {
 	id: string;
 	title: string;
 	description: string;
 	area: string;
 	blockTemplates: Template[];
-};
-
-export type ProductTemplate = {
-	id: string;
-	title: string;
-	description: string | null;
-	icon: string | null;
-	order: number;
-	layoutTemplateId: string;
-	productData: Partial< Product >;
 };
 
 export type ProductEditorSettings = Partial<
